+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # RS3.0
-
-Compatible with disguise designer version r31.1 and above.
+Compatible with disguise designer version r34.0 and above.
 * Added project working colour space in schema
+* When using texture parameters, the engine is required to call `rs_registerTextureParams` to register the received frames before they are used. This must be called from a render thread.
 
 # RS2.0
 Compatible with disguise designer version r25.0 and above.